Perform Repeated Rarefaction and Generate an Ordination Plot using PCoA

This function performs repeated rarefaction on a phyloseq object, computes ordination, and generates a PCoA-based visualization. The same procedure is used from threshold_testing function when testing a range of thresholds.

Usage

repeated_rarefaction(
  input,
  repeats = 50,
  threshold = 250,
  colorb = "sample_id",
  group = "sample_id",
  cloud = TRUE,
  ellipse = FALSE,
  cores = 2,
  ...
)

Arguments

input

A phyloseq object.

repeats

An integer. The number of times to repeat rarefaction. A value of 1 means no repeats. Default = 50.

threshold

An integer. The threshold value to use for rarefaction. Default = 250.

colorb

A string. Column name in the sample_data() bundled with the supplied phyloseq object, used to color sample points.

group

A string. Column name in sample_data() used for grouping the samples. The parameter is also used to calculate an ellipse around the points. The supplied value should coincide with the group when using the function threshold_testing, since it is on this parameter that also the clustering performance index is calculated.

cloud

A boolean. If TRUE, all repeated data points are shown. Otherwise, only the median points of each sample repetitions cloud are plotted. Default = FALSE.

ellipse

A boolean. If TRUE, confidence ellipses around samples are from the same group are drawn. Default = TRUE.

cores

An integer. Number of cores to use for parallel processing. Default = 2.

...

Additional arguments to be passed to the function.

Value

A list containing (While also showing the plot directly):

repeats: Number of repeats.

df_consensus_coordinates: Median ordination positions.

df_all: All ordination positions.

plot: a ggplot object.