Test Different Rarefaction Thresholds Using Repeated Rarefaction

This function is a wrapper around the steps of repeated_rarefaction that evaluates different rarefaction thresholds. It runs repeated rarefactions on a phyloseq object and calculates clustering performance using the Calinski-Harabasz index. Moreover, the function also calculates the average pairwise distance for each sample across different rarefaction thresholds, to give a measure of sample concordance.

Usage

test_threshold(
  input,
  repeats = 50,
  t_min = 50,
  t_max = 250,
  t_step = 5,
  group = "sample_id",
  cores = 2,
  ...
)

Arguments

input

A phyloseq object.

repeats

An integer or a vector of integers. The number of times to repeat rarefaction. A value of 1 means no repeats. If using a vector, different rarefaction thresholds will be tested sequentially. Default = 50.

t_min

An integer. The minimum value for the threshold testing range. Default = 50.

t_max

An integer. The maximum value for the threshold testing range. Default = 250.

t_step

A numeric value. The step size for the threshold testing range. A value between 0 and 1 will cause the same threshold to be tested multiple times. Default = 5.

group

A string. A string. Column name in sample_data() used for grouping the samples. It is also on this value that the Calinski-Harabasz index is calculated.

cores

An integer. Number of cores for parallel processing. Default = 2.

...

Additional arguments to be passed to the function.

Value

A list containing:

index_plot: A ggplot object showing Calinski-Harabasz index vs. rarefaction threshold.

ordination_plots: A list containing the ordination plots for tested threshold values. The ordination plots have different levels corresponding to the number of repeats.

avg_distances: A data frame (or list of data frames if using multiple repeat values) with average pairwise distances of the cloud generated from each sample repetition attempts (for each threshold).

Details

Higher index values are better, and the plateauing of the index values when testing higher thresholds indicate that the clouds of repetitions are compact enough to not being affected by increasing the threshold. The Calinski-Harabasz value takes into account both within cluster and between cluster distances, and in the scope of this function is calculated on the group parameter. A sudden drop in CH value means that samples might have been removed because they did not have enough reads to reach the threshold. In this case a warning is printed listing the samples which have been removed. In avg_distances, if samples do not have enough reads to reach t_max value, their APD are set to 0 after that threshold and a warning message is printed.